Code refactoring after PR review
