+++ page_objects/pageManager.ts
@@ -6,6 +6,11 @@ import { VeterinariansPage } from './veterinariansPage';
 import { PetTypesPage } from './petTypesPage';
 import { PetDetailsPage } from './petDetailsPage';
 import { SpecialitiesPage } from './specialitiesPage';
+import { EditVeterinarianPage } from './editVeterinarianPage'; 
+import { NewVisitPage } from './newVisitPage'; 
+import { EditPetTypePage } from './editPetTypePage';
+import { EditSpecialityPage } from './editSpecialityPage';
+
 
 
 export class PageManager{
@@ -18,6 +23,11 @@ export class PageManager{
     private readonly ownersPage: OwnersPage
     private readonly ownerInformationPage: OwnerInformationPage
     private readonly specialitiesPage: SpecialitiesPage
+    private readonly editVeterinarianPage: EditVeterinarianPage
+    private readonly newVisitPage: NewVisitPage
+    private readonly editPetTypePage: EditPetTypePage
+    private readonly editSpecialityPage: EditSpecialityPage
+
 
     constructor(page: Page){
         this.page = page
@@ -28,6 +38,11 @@ export class PageManager{
         this.ownersPage = new OwnersPage(this.page)
         this.ownerInformationPage = new OwnerInformationPage(this.page)
         this.specialitiesPage = new SpecialitiesPage(this.page)
+        this.editVeterinarianPage = new EditVeterinarianPage(this.page)
+        this.newVisitPage = new NewVisitPage(this.page)
+        this.editPetTypePage = new EditPetTypePage(this.page)
+        this.editSpecialityPage = new EditSpecialityPage(this.page)
+        
     }
 
     navigateTo(){
@@ -54,7 +69,23 @@ export class PageManager{
         return this.ownerInformationPage
     }
 
-    onMainSpecialitiesPage(){
+    onSpecialitiesPage(){
         return this.specialitiesPage
+    }
+
+    onEditVeterinarianPage(){
+        return this.editVeterinarianPage
+    }
+
+    onNewVisitPage(){
+        return this.newVisitPage
+    }
+
+    onEditPetTypePage(){
+        return this.editPetTypePage
+    }
+
+    onEditSpecialityPage(){
+        return this.editSpecialityPage
     }
 }
