+++ page_objects/specialitiesPage.ts
@@ -6,22 +6,11 @@ export class SpecialitiesPage {
         this.page = page
     }
 
-    /**
-     * This method updates speciality, availiable in the main Specialities list
-     * @param speciality - expected speciality
-     */
-    async editMainSpecialityTo(speciality: string){
-        const specialityInputField = this.page.getByRole('textbox')
-        await specialityInputField.click()
-        await specialityInputField.clear()
-        await specialityInputField.fill(speciality)
-    }
-
-    async deleteSpelialityByName(speciality: string){
+    async deleteSpecialityByName(speciality: string){
         await this.page.getByRole('row', {name: speciality}).getByRole('button', {name:"Delete"}).click()
     }
 
-    async addNewMainSpeciality(newSpeciality: string){
+    async addNewSpeciality(newSpeciality: string){
         await this.page.getByRole('button', {name:"Add"}).click();
         const addSpecialityField = this.page.getByRole('textbox').last()
         await addSpecialityField.click()
@@ -33,22 +22,22 @@ export class SpecialitiesPage {
      * Extrats all specialities text from Specialities table.
      * @returns a list of strings with speciality names from Specialities table.
      */
-    async createListOfAllMainSpecialitiesTable(){
+    async createListOfAllSpecialitiesThatAreShownInTable(){
         const allRows = this.page.getByRole('row').filter({has: this.page.getByRole('button', {name:"Edit"})})
-        let allMainSpecialitiesList: string[] = [];
+        let allSpecialitiesList: string[] = [];
 
         for(let row of await allRows.all()){
             let spec = await row.locator('td input').inputValue()
-            allMainSpecialitiesList.push(spec)
+            allSpecialitiesList.push(spec)
         }
-        return allMainSpecialitiesList
+        return allSpecialitiesList
     }
 
     /**
      * Extrats all specialities options text from Specialities dropdown.
      * @returns a list of strings with speciality names from Specialities dropdown.
      */
-    async createListOfAllMainSpecialitiesInDropdownOptions(){
+    async createListOfAllSpecialitiesInDropdownOptions(){
         await this.page.locator('.dropdown-display').click()
 
         const dropdownOptions = this.page.locator('.dropdown-content div label')
@@ -72,8 +61,6 @@ export class SpecialitiesPage {
         await expect(selectedSpecialtyRow.locator('td input')).toHaveValue(expectedSpeciality);
     }
 
-    async confirmSpecialityUpdate(){
-        await this.page.getByRole('button', { name: 'Update' }).click()
-    }
+   
 }
